% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffymetrixCelSetList.R
\name{AffymetrixCelSetList}
\alias{AffymetrixCelSetList}
\title{Make a list object where each element is an AffymetrixCelSet}
\usage{
AffymetrixCelSetList(celSet, type, tags, path = NULL)
}
\arguments{
\item{celSet}{The parent celSet from which the stored parameters are derived}

\item{type}{The type of values stored. 
Can take the values "backgroundPriors", "contrastData" or"modelData"}

\item{tags}{Specifies which tags have been added to the parentName.
See Details for more information}

\item{path}{Where to look for the data}
}
\value{
A list of AffymetrixCelSet objects
}
\description{
Used for storing values for parameters and contrasts under the BMEA model
}
\details{
For each type of celSetList, only a limited number of tags are valid:
\itemize{
\item{type=backgroundPriors}{tags can be QN, MAT, GC, RMA, GC-RMA or the like}
\item{type=contrastData}{tags can only be logFC of phiLogFC}
\item{type=modelData}{tags can be any of the model parameter names (i.e. S, sigmaS, c, mu, sigmaMu, p, sigmaP, phi)}
}

By default, all files found with the correct specification will be included in the list object

The cdf will be assigned automatically, 
i.e. BG lists will use the same as the celSet, 
contrasts will use the monocell CDF & model parameters will use the full (S & p) or monocell (sigmaS, c, mu, sigmaMu, sigmaP, phi) respectively
}

