% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotPhi.R
\name{boxplotPhi}
\alias{boxplotPhi}
\title{Draws a boxplot of the phi distribution for all exons/groups.}
\usage{
boxplotPhi(data, unit, which = 1, maxB = log(6001),
  bThresh = log(2997.5/3.5), nProbes, grpNames, featName = NULL,
  showGroups = TRUE, unchCol = "white", upCol = "green",
  downCol = "red", ..., widths, main = NULL, medLwd = 2, lty = 1,
  las = 2, cex.x = 0.9, cex.y = 0.9, h = 0, abCol = "blue",
  hLab = "*", col.axis = c("black", "grey"))
}
\arguments{
\item{data}{can be either an \code{array}, \code{matrix} or \code{\link{AffymetrixCelSetList}}}

\item{unit}{specifies which unit to plot if \code{data} is an \code{\link{AffymetrixCelSetList}}.
Ignored if \code{data} is an \code{array} or \code{matrix}.}

\item{which}{specifies which condition/cell-type or contrast to plot. 
Defaults to the first condition or contrast}

\item{maxB}{the maximum possible B-statistic from the entire dataset. 
Should be \code{log(2*(nChains*(nIter-nBurnin))/nThin+1)} from the \code{mcmcParam} object in the general workspace. 
Only used if plotting a contrast.}

\item{bThresh}{the maximum B-statistic to display in colour}

\item{nProbes}{the number of probes per group. 
Will over-ride any values found on an associated \code{AffymetrixCdfFile} for a given unit}

\item{grpNames}{the names of the group-level probesets. 
Will over-ride any found on an \code{AffymetrixCdfFile}}

\item{featName}{the feature being plotted. 
Should be a cell-type/condition if plotting raw phi values, or the contrast name if plotting a contrast. 
Not required if extracting data from an \code{\link{AffymetrixCelSetList}}}

\item{showGroups}{logical. Determines whether to diplay the groupNames on the x-axis.}

\item{unchCol}{the default fill colour for the boxplots}

\item{upCol}{the colour to use for contrast values greater than \code{bThresh}}

\item{downCol}{the colour to use for contrast values less than  -\code{bThresh}}

\item{...}{passes any further specified graphical parameters}

\item{widths}{the width of the boxes. 
Over-rides any values calculated from an \code{AffymetrixCdfFile}}

\item{main}{the title of the plot}

\item{medLwd}{the width of the line to draw for the medians}

\item{lty}{the line-type for the whiskers}

\item{las}{see \code{\link{plot}}}

\item{cex.x}{\code{cex.axis} for the x-axis labels}

\item{cex.y}{\code{cex.axis} for the y-axis labels}

\item{h}{values at which a horizontal line will be drawn}

\item{abCol}{the colour of any horizontal-lines}

\item{hLab}{text characters to denote significantly up/down changed  groups}

\item{col.axis}{the colour for the main axes & any zero line}
}
\description{
Draws a boxplot of the posterior phi distributions for all exons within a transcript.
Can be either a contrast (phiLogFC) or a parameter (phi).
}
\details{
Plots the posterior distributions for the exon proportions (phi) as specified by the BMEA model. 
Either the distributions for a given cell type, or a given contrast will be plotted.

Whiskers will extend from the 75th to 97.5th quantile and from the 25th to the 2.5th quantile. 
The box will display the IQR.

In the case of a contrast, any significant groups will be indicated by default with an asterisk. 
These will only appear for groups with a B-statistic equal to the maximum possible value. 
Groups with B-statistics less than this value, but above the specified value for \code{bThresh} will be coloured using \code{upCol} or \code{downCol}.
}
\seealso{
\code{\link{AffymetrixCelSetList}}, \code{\link{AffymetrixCdfFile}}, \code{\link{boxplot}}
}

