% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearNodes.R
\name{clearNodes}
\alias{clearNodes}
\title{Remove all files/folders used on each node}
\usage{
clearNodes(nodePaths)
}
\arguments{
\item{nodePaths}{a list where each element contains the folder 
where the node-specific CEL & log files are stored}
}
\description{
After running the BMEA process in parallel, 
this function removes all the node-specific files
}
\details{
NB: This must only be run after the nodes have been successfully merged,
as all information saved in the node folders will be deleted permanently
}
