% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCelHeader.R
\name{createCelHeader}
\alias{createCelHeader}
\title{Create a CEL header from the supplied inputs}
\usage{
createCelHeader(filename = NULL, cdf = NULL, version = 4,
  cols = NULL, rows = NULL, algorithm = "Percentile",
  parameters = NULL, chiptype = NULL, header = NULL,
  datheader = NULL, librarypackage = "", cellmargin = 4,
  noutliers = 0, nmasked = 0, tags = NULL)
}
\arguments{
\item{filename}{the name of the CEL file that the header will be written to}

\item{cdf}{the cdf for the CEL file}

\item{version}{the version number of the CEL file to be created}

\item{cols}{the number of columns for the CEL file. 
Ignored if a \code{cdf} is supplied}

\item{rows}{the number of rows for the CEL file. 
Ignored if a \code{cdf} is supplied}

\item{algorithm}{a text string to be written to the \code{$algorithm} component}

\item{parameters}{a text string to be written to the \code{$parameters} component}

\item{chiptype}{a text string defining the chiptype to be  written. 
Ignored if a \code{cdf} is supplied}

\item{header}{a text string to be written to the \code{$header} component. 
Must be in the correct format}

\item{datheader}{a text string to be written to the \code{$datheader} component. 
Must be in the correct format}

\item{librarypackage}{a text string to be written to the \code{$librarypackage} component. 
Defaults to ""}

\item{cellmargin}{a text string to be written to the \code{cellmargin} component}

\item{noutliers}{the value to be written to the \code{$noutliers} component}

\item{nmasked}{defaults to zero}

\item{tags}{An optional text string which will be placed before the filename in the \code{$datheader} component}
}
\value{
Returns a list with the following components:
\itemize{
\item{$filename}{ the filename of the CEL file to be created}
\item{$version}{ the version of the CEL file}
\item{$cols}{ the number of rows to be generated on the CEL file}
\item{$rows}{ the number of rows to be generated on the CEL file}
\item{$total}{ the total number of cells. Will always equal \code{$rows}*\code{$cols}}
\item{$algorithm}{ a typical algorithm which is available for scanners}
\item{$parameters}{ a typical list of parameters as would be found on a CEL file}
\item{$chiptype}{ the type of CEL file to be written}
\item{$header}{ the header in the correct format}
\item{$datheader}{ a typical header output from a .DAT file}
\item{$librarypackage}{ can be any text string}
\item{$cellmargin}{ an integer}
\item{$noutliers}{ an integer}
\item{$nmasked}{ an integer}
}
}
\description{
Create a list in the same format as a CEL header, 
in preparation for creating a CEL file
}
\details{
This function will output a list in the correct format for generation of a blank CEL file using \code{createCel}.

Many of the input values are redundant, but must be included to pass checks during the creation of CEL files.

If a cdf is specified, any values for \code{cols}, \code{rows} or \code{chiptype} will be ignored, and those on the cdf will be used.
}
\seealso{
\code{\link{createCel}} \code{\link{AffymetrixCdfFile}}
}
