% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineGcBins.R
\name{defineGcBins}
\alias{defineGcBins}
\title{Estimate the mean and standard deviation for the log background signal}
\usage{
defineGcBins(celSet, cdf = NULL, bgProbes = "r2.genomic.bgp", nBins = 20,
  fitIndividual = TRUE)
}
\arguments{
\item{celSet}{the set of arrays being processed, as an \code{AffymetrixCelSet}.}

\item{cdf}{an alternate cdf. 
This will override the one specified in \code{celSet}. 
If using a custom Affymetrix cdf, 
the original Affymetrix cdf should be specified here as the custom one may not contain the set of background probes.}

\item{bgProbes}{the tab delimited file containing the sequence data for each background probe.}

\item{nBins}{the number of bins to form. Defaults to 20}

\item{fitIndividual}{logical variable. 
If \code{fitIndividual=TRUE}, the parameters will be estimated on an array-specific basis, 
otherwise the parameters will be estimated across the entire celSet.}
}
\value{
A list with the following components:
  \itemize{ 
  \item{lambda}{the mean of the log-transformed \code{bgProbes} for a given \code{gc_count}}
  \item{delta}{the standard deviation of the log-transformed \code{bgProbes} for a given \code{gc_count}}
  }
  For easy downstream usage, the names/rownames for each component are specified as "GC4" through to "GC25."
}
\description{
Based on the supplied set of background probes, 
estimate the background mean and standard deviation for a given GC count.
}
\details{
This function finds the log scale mean (lambda) and standard deviation (delta) of the specified \code{bgProbes} for each \code{gc_count}.
}

