% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMatBins.R
\name{defineMatBins}
\alias{defineMatBins}
\title{Define the bins, mean & sd for log background signal using the MAT model.}
\usage{
defineMatBins(bgParam, nBins = 20, fitIndividual = TRUE)
}
\arguments{
\item{bgParam}{the output from the function \code{fitBackgroundParameters}. 
Must contain the slot \code{$coef}.}

\item{nBins}{specifies the number of bins to form based on the fitted MAT values. 
Defaults to 20 bins.}

\item{fitIndividual}{logical.Determines whether bin means & sds are estimated for each array or for the entire dataset.}
}
\value{
Returns a list with the components: 
  \itemize{
  \item{$boundaries}{the boundaries for the bins. The lower boundary is implied as zero & 
  the upper boundary is implies as \code{Inf}.}
  \item{$lambda}{the observed mean of the background probes classified into each bin}
  \item{$delta}{ the observed standard deviation of the background probes as classified into each bin}
  }
  
  All means and standard deviations are on the log-scale and refer to the normal distributions for log(Background).
}
\description{
Define the boundaries for the background signal, 
based on the fitted MAT values for each probe, and the subsequent means & standard deviations.
}
\details{
Given the fitted MAT values from the background set of probes in the \code{bgParam} argument, 
the boundaries for forming the background prior distributions are calculated using this function. 
Bins are formed using approximately equal numbers of probes.
}
\references{
Kapur, K., Xing, Y., Ouyang, Z., Wong, WH. (2007) \emph{Exon arrays provide accurate assessments of gene expression} 
 Genome Biol. 8(5):R82 
  
 Johnson, W.E., Li, W., Meyer, C.A., Gottardo, R., Carroll, J.S., Brown, M., Liu, X.S. (2006) \emph{Model-based analysis of tiling-arrays for ChIP-chip.}
 Proc Natl Acad Sci USA 103:12457-12462
}
\seealso{
\code{\link{fitBackgroundParameters}}
}

