% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBmeaArray.R
\name{extractBmeaArray}
\alias{extractBmeaArray}
\title{Extract the data from an AffymetrixCelSetList as an array or matrix}
\usage{
extractBmeaArray(celSetList, units = NULL, which = NULL,
  firstOnly = TRUE, log = TRUE)
}
\arguments{
\item{celSetList}{an object of class \code{\link{AffymetrixCelSetList}}}

\item{units}{the units to be extracted. Can be either a numeric vector or
a character string with the unitNames as on the appropriate cdf. 
If not supplied, all units will be returned.}

\item{which}{specify which elements of the celSetList should have data extracted from them. 
If not supplied, data from all elements of the celSetList will be returned}

\item{firstOnly}{if \code{firstOnly=TRUE}, only the first cell from each unit will be extracted. 
If \code{firstOnly=FALSE} all cells from each unit will be extracted.}

\item{log}{should the data be log-transformed}
}
\value{
Returns either a matrix or array, depending on which data is requested
}
\description{
Extract the data saved to an AffymetrixCelSetList for a given set of units
}
\details{
During the BMEA process, contrast, model parameters & background priors are all saved as AffymetrixCel Files. 
This enables simple extraction of saved data for any combination of units & files.
}
