% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBackgroundParameters.R
\name{fitBackgroundParameters}
\alias{fitBackgroundParameters}
\title{Estimate the parameters for the Background Signal.}
\usage{
fitBackgroundParameters(celSet, cdf = NULL,
  bgProbes = "r2.genomic.bgp", method = "MAT")
}
\arguments{
\item{celSet}{the set of .CEL files being analysed as an \code{AffymetrixCelSet}}

\item{cdf}{the cdf required for accessing the background probe intensities. 
Some custom cdf files do not contain any background probes.}

\item{bgProbes}{the tab delimited file containing the sequence data for each background probe.}

\item{method}{specifies the model for background signal. 
Currently only "MAT" is implemented.}
}
\value{
Returns a list with the following components:
\itemize{
 \item{$coef}{ the fitted model coefficients}
 \item{$fitted}{ the fitted values for each background probe}
 \item{$observed}{ the observed values for each background probe}
 \item{$method}{ a character string denoting which method was used for model fitting.}
 \item{$chipType}{ a character string denoting which chipType (& cdf) was used for fitting the model.}
}
}
\description{
Fits the background signal model for a set of exon arrays
}
\details{
This function fits the model parameters for background signal estimates, using the set of probes specified in the 'bgProbes' argument. 
Currently only the modified MAT model has been implemented.

This function is dependent on the file directory structure as used by aroma.affymetrix. 
The .CEL files must be placed in the \code{.../probeData/celSetName/chipType} directory, 
and supplied to the function as an \code{AffymetrixCelSet}. 

Quantile normalisation & optical background correction prior to processing with this function 
is optional but recommended.

The set of background probes to be used for fitting the model parameters must be located in the \code{.../annotationData/chipTypes/chipType} directory. 
The filename must begin with the chipType & end with the suffix as supplied in the bgProbes argument, e.g. "HuEx-1_0-st-v2.r2.genomic.bgp". 
The file format must be the same as the .bgp files as supplied by affymetrix, i.e. must have 10 columns, of which one is "probe_sequence".
}
\references{
H. Bengtsson, K. Simpson, J. Bullard, and K. Hansen, (2008) 
\emph{aroma.affymetrix: A generic framework in R for analyzing small to very large Affymetrix data sets in bounded memory}, 
Tech Report #745, Department of Statistics, University of California, Berkeley. 

Kapur, K., Xing, Y., Ouyang, Z., Wong, WH. (2007) 
\emph{Exon arrays  provide accurate assessments of gene expression} 
Genome Biol. 8(5):R82

Johnson, W.E., Li, W., Meyer, C.A., Gottardo, R., Carroll, J.S.,Brown, M., Liu, X.S. (2006) 
\emph{Model-based analysis of tiling-arrays for ChIP-chip.}
Proc Natl Acad Sci USA 103:12457-12462
}
