% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBmea.Snow.R
\name{fitBmea.Snow}
\alias{fitBmea.Snow}
\title{Fit the BMEA process for a number of genes utilising the snow package}
\usage{
fitBmea.Snow(celSet, bgCelSet, cl, units = NULL, batchSize = NULL,
  conditions, contMatrix, ..., nodePaths = NULL)
}
\value{
Returns a list object with the elements:
\itemize{
\item{$units}{ a list with each sub-element containing the units fitted on each node. 
Each sub element is named with the provided \code{nodePath}}
\item{$times}{ a list with the times each batch was written to disk for each node.}
\item{$parameters}{ the model parameters that have been saved on each node}
}
}
\description{
Enables genes to be fit in parallel, utilising the architecture of the snow package
}
\details{
This process will split the supplied units into vectors of equal length, 
with one for each node in the supplied cluster. 
The BMEA process will be fit in series for the vector of units on each parallel node. 
Depending on the computing architecture being used & the size of the dataset, 
this process can still take a number of hours, to a number of days.

Due to the architecture of the \code{snow} package, all arguments will be first sent to
.GlobalEnv with the prefix 'bmeaTemp.'. 
Any objects in .GlobalEnv with this prefix will be deleted as part of the process, 
so users should avoid using this string as part of the name of any objects.
 
On each node, the genes will be fit in series & the results saved to disk periodically as a batch,
with the results saved as CEL files. 
The number of genes fit in each batch is determined by the \code{batchSize} argument. 
Each time the results for a batch of units are written to disk, 
a log file with the name \code{parentName.log.txt} will be updated to ensure minimum data is lost in the case of any interuptions. 
These log files are recorded in a node-specific manner in the directory "nodePath/bmeaData". 
The results for each parameter in the \code{paramToSave} argument will be written to CEL files 
on each node in the directories \code{nodePath/bmeaData/modelData}, 
whilst the contrasts specified in the contrast matrix will be written to the directories \code{nodePath/bmeaData/contrastData}.

If the \code{mcmcParam} argument is not specified as part of the '...' the defaults of \code{nChains=3},
\code{nIter=16000}, \code{nBurnin=8000} & \code{nThin=8} will be used. 
The default values for \code{zGene} (4.265) & \code{zExon} (1.645) will also be used if not specified.

If the \code{paramToSave} argument is not specified, the default of the chip effects ("c"), 
condition-specific effects ("mu") and the exon-level terms ("phi") will be saved. 
If the \code{paramToWrite} argument is not specified, it will default to the same as the \code{paramToSave} argument.

After the process has been run, the saved data for the sets of parallel units can be merged via the function \code{\link{mergeNodes}}.
}
\seealso{
\code{\link{mergeNodes}}, \code{\link{AffymetrixCelSet}}, 
 \code{\link{AffymetrixCelSetList}}, \code{\link{makeCluster}}
}

