% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBmeaBatch.R
\name{fitBmeaBatch}
\alias{fitBmeaBatch}
\title{Fit the BMEA model for a single batch of units}
\usage{
fitBmeaBatch(celSet, bgCelSet, units, conditions, contMatrix, ...,
  paramToSave = c("c", "mu", "phi"), keepSims = FALSE, zGene = 4.265,
  zExon = 1.645)
}
\arguments{
\item{celSet}{an \code{AffymetrixCelSet} with the data to be fit}

\item{bgCelSet}{a list with components \code{$lambda} & \code{$delta}. 
Each of these must be an \code{AffymetrixCelSet} containing the means & standard deviations
for the background signal priors}

\item{units}{the units (i.e. genes) to be fit}

\item{conditions}{a vector of factors specifying which cell-type/condition each array in the \code{celSet} belongs to}

\item{contMatrix}{a contrast matrix for the summarised output}

\item{...}{used for passing further arguments such as \code{mcmcParam} to \code{runMCMC.BMEA}}

\item{paramToSave}{the model parameters to be saved for downstream processing. 
The parameters "c", "mu" & "phi" will always be saved.}

\item{keepSims}{logical variable. 
If \code{TRUE} all sims from the process & contrasts will be kept}

\item{zGene}{the zScore below which a gene is classified as not detectable above background}

\item{zExon}{the zScore below which an exon is classified as not detectable above background}
}
\value{
An object of class("BMEA.Batch"), which is a list with the following components:
 
 \itemize{
  \item{$celSet}{ the \code{celSet} being analysed, as supplied to the function}
  \item{$summaries}{ a \code{list} with a component for each unit. 
  Each component contains the summary statistics for the unit, including the convergence statistics "rHat" & "nEff."}
  \item{$logFC}{ a \code{list} with a component for each contrast supplied in \code{contMatrix}. 
  Each row contains the summary statistics for a single unit, for that contrast}
  \item{$phiLogFC}{ a \code{list} with a component for each contrast supplied in \code{contMatrix}. 
  Each row represents an exon (group).}
  \item{$conditions}{ the cell-types (or conditions) as factors, as supplied to the function}
  \item{$units}{ a \code{data.frame} with the units fit & the  corresponding unitNames.}
  \item{$paramToSave}{ the parameters requested to be saved.}
  \item{$sims}{ a \code{list} with a component for each unit. 
  If \code{keepSims=FALSE}, will return \code{NULL} for each component.}
  }
}
\description{
Fits the BMEA model sequentially for more than one unit
}
\details{
This is the function used to fit the BMEA model to a batch of units (or genes). 
Each unit is tested to see if it contains multiple exons, and is expressed detectably 
above background before analysis. 
For single exon genes, all exon-level terms are omitted from the model, 
as the PLM model used for conventional 3' Arrays holds for these genes & can be used
with minimal computational effort.

Units that are not fitted are also removed from the output vector of units.

Restricting the parameters to be saved, via the \code{paramToSave} argument can significantly 
save the memory requirements for large batches of genes. 
This will default to the parameters "c", "mu" & "phi". 
The signal parameter "S" is the most demanding on memory resources &
is generally not advised to be saved unless it is of specific interest.
}
\seealso{
\code{\link{fitBmeaSingle}}, \code{\link{writeBmeaBatch}}
}

