% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcCounts.R
\name{gcCounts}
\alias{gcCounts}
\title{Get the number of GC bases from a batch of sequences.}
\usage{
gcCounts(seq)
}
\arguments{
\item{seq}{a vector of 25bp sequences}
}
\value{
Returns a \code{vector} the same length as the input \code{vector} of sequences
}
\description{
Counts the number of G & C bases in a character vector of sequences
}
\details{
Counts the number of G & C bases in a batch of a sequences
}
