% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogFC.R
\name{getLogFC}
\alias{getLogFC}
\title{Sample the log fold-change from a BMEA.MCMC dataset}
\usage{
getLogFC(data, contr.matrix, keepSims = FALSE)
}
\arguments{
\item{data}{an object of class BMEA.MCMC. 
Will contain a set of simulations for every condition-specific expression level.}

\item{contr.matrix}{a contrast matrix with rows representing each condition & columns 
representing the specific contrasts. 
The column names are recycled as rownames in the summary output & for thesimulations if kept.}

\item{keepSims}{logical variable. 
Determines whether the sampled values are returned in the output or just the summary statistics. 
Defaults to FALSE.}
}
\value{
A list with the following components:
\itemize{
\item{summary}{a summary matrix with mean, standard deviation and the quantiles for each contrast. 
Each row represents a different contrast.}
\item{sims}{a matrix of the sampled posterior disitributions for each contrast. 
Rows represent a kept iteration from the MCMC process, and columns represent each contrast. 
Returns NULL if keepSims is set to FALSE}
}
}
\description{
Samples the posterior distribution of log fold-change for each specified contrast from the given dataset.
}
\details{
Obtains the posterior distribution for log fold-change for each contrast by direct sampling from 
the MCMC output. 
A summary for each contrast with mean, sd & the quantiles are returned in the $summary argument
as the default.

All data is retrieved using C and this function is a simple wrapper to the C function.
}

