% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBMEA.R
\name{isBMEA}
\alias{isBMEA}
\alias{isBMEA.Batch}
\alias{isBMEA.MCMC}
\title{Check the class & structure of BMEA objects}
\usage{
isBMEA(x)

isBMEA.Batch(x)

isBMEA.MCMC(x)
}
\arguments{
\item{x}{the object to be tested}
}
\value{
logical
}
\description{
Check the class & structure of BMEA objects. Generally for internal usage
}
\details{
If an object is required for downstream analysis, this functions check the class attribute & the structure of the object to ensure forward compatability.
}
