% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNodes.R
\name{mergeNodes}
\alias{mergeNodes}
\title{Merge the CEL files written on multiple nodes into a single set}
\usage{
mergeNodes(celSet, nodePaths, outDir = NULL, ..., paramToWrite = NULL,
  nodeUnits = NULL, visible = FALSE, verbose = TRUE)
}
\arguments{
\item{celSet}{the parent celSet being analysed}

\item{nodePaths}{where to find all the node specific CEL files}

\item{outDir}{the directory to merge the node-level information to.}

\item{...}{not used}

\item{paramToWrite}{the model parameters to collect from each node & write to the merged CEL files}

\item{nodeUnits}{a \code{\link{list}} with a vector of units for each node that correspond to those processed on each node. Overrides any information contained in the log files. 
The units will be extracted from the log files if this is not supplied}

\item{visible}{if \code{visible=TRUE} a \code{\link{list}} will be returned, 
otherwise the function will invisibly return \code{TRUE}. Defaults to FALSE}

\item{verbose}{determines the amount of information to display. Defaults to \code{TRUE}}
}
\value{
If \code{visible=FALSE} will invisibly return \code{TRUE}, otherwise a \code{\link{list}} will be 
returned with the following elements:
\itemize{
\item{$written}{ \code{TRUE}}
\item{$parameters}{ contains the model parameters that have been written to the set of CEL files.}
\item{$nodeUnits}{ the units processed on each node}
\item{$nodePaths}{ the paths the node specific CEL files were sourced from}
\item{$logFiles}{ the contents of the log files from each node}
}
}
\description{
Takes the node specific sets of CEL files containing BMEA contrast & 
model parameter data and merges them into a single set of CEL files.
}
\details{
During analysis using parallel nodes, 
each node will write to a batch of CEL files in its own separate directory. 
This function collects the units that have been run on each node and copies the saved results 
from the CEL files on each node to a main set of CEL files in the root directory.

CEL files with the contrast data (i.e. logFC & phiLogFC) will be written to the directories 
\code{bmeaData/contrastData/parentName,logFC/chipType} & \code{bmeaData/contrastData/parentName,phiLogFC/chipType} respectively. 
If no output directory is set via \code{outDir} the current working directory will be used.

CEL files with model parameters will be written to the directories \code{bmeaData/modelData/parentName,parameterName/chipType}.
}
\seealso{
\code{\link{AffymetrixCelSet}}
}

