% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMC.BMEA.R
\name{runMCMC.BMEA}
\alias{runMCMC.BMEA}
\title{Fit the BMEA model for a single gene.}
\usage{
runMCMC.BMEA(PM, conditions, exons, lambda, delta, mcmcParam = NULL,
  model = "Full", ...)
}
\arguments{
\item{PM}{the raw intensity data for a single gene.}

\item{conditions}{a vector of factors defining which array belongs to which cell-type, or condition.}

\item{exons}{a vector of factors defining which probe belongs to which exon (or group).}

\item{lambda}{a matrix of the same dimensions as the PM matrix. 
Contains the probe-specific means for the log-normal distributions used as the priors for the background signal.}

\item{delta}{a matrix of the smae dimensions as the PM matrix. 
Contains the standard deviations for the log-normal distributions used as the priors for the background signal.}

\item{mcmcParam}{a list defining the parameters for the MCMC process. 
Requires the arguments: 
 \itemize{
 \item{$nChains}{ the number of chains to run. Defaults to 3 if not specified.}
 \item{$nIter}{ the number of iterations to run. Defaults to 12,000 if not specified.}
 \item{$nBurnin}{ the number of burnin iterations. Defaults to the first half of the iterations.}
 \item{$nThin}{ the number of iterations used to thin the stored data, 
 e.g. if \code{$nThin} = 10, every 10th iteration will be kept. Defaults to 6 if not specified.}
 }}

\item{model}{a character string defining which BMEA model to run. 
Currently only the full model (\code{model="Full"}) or one with no probeEffects 
(\code{model="noProbes"}) are implemented.}

\item{...}{not used}
}
\value{
Returns a 3D-array of class BMEA.MCMC. 
The first dimension contains the model parameters for the given gene. 
The second dimension stores the iterations for each chain, 
whilst the third dimension contains the number of chains used for the process.

The names for the first dimension are vital for downstream analysis &  
will specify which parameters are mu & phi.
}
\description{
Fits the BMEA model using the MCMC process for a single gene.
}
\details{
This function runs the complete BMEA process for a single gene. 
Currently the only implemented model is the uniform prior for the exon proportions term (phi). 

Computation times will vary to a large degree based on the processor involved. 
It should be noted that the process for a single gene can take in the order of minutes. 

The MCMC process itself is wrtten in C and this function is a wrapper to this function.
}

