% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMatMatrix.R
\name{setMatMatrix}
\alias{setMatMatrix}
\alias{setMatVector}
\title{Define the MAT model matrix for an given set of sequences}
\usage{
setMatMatrix(seq)

setMatVector(seq)
}
\arguments{
\item{seq}{genomic sequences. \code{setMatVector} only takes a single string, 
whilst \code{setMatMatrix} takes a vector of strings.
Each sequence must be a string of 25 characters containing only the letters A, C, G & T
 
The function setMatVector is designed for one individual sequence.}
}
\value{
\code{setMatMatrix} Returns a matrix with nrows=length(seq) & ncol = 80.

\code{setMatVector} nReturns a vector of length 80
}
\description{
Defines the MAT model matrix for a set of sequences.
}
\details{
Calculates the model matrix for the MAT background model for a given set of genomic sequences. 
Allows fitting of the MAT model for the purposes of background signal estimation.

setMatVector is designed for internal use only.
}
\examples{
## Generate a random sequence
seq <- sample(c("A","C","G","T"), 25, replace=TRUE)
seq <- paste(seq,collapse="")
mat.vec <- setMatVector(seq) ## Returns a vector of length 80

## Generate a random sequence
seq <- replicate(2, sample(c("A","C","G","T"), 25, replace=TRUE), simplify = FALSE)
seq <- sapply(seq, paste, collapse = "")
mat.matrix <- setMatMatrix(seq) ## Returns a 2x80 matrix

}
\references{
Kapur, K., Xing, Y., Ouyang, Z., Wong, WH. (2007) 
  \emph{Exon arrays provide accurate assessments of gene expression} 
  Genome Biol. 8(5):R82
  
  Johnson, W.E., Li, W., Meyer, C.A., Gottardo, R., Carroll, J.S., Brown, M., Liu, X.S. (2006) 
  \emph{ Model-based analysis of tiling-arrays for ChIP-chip.}
  Proc Natl Acad Sci USA 103:12457-12462
}
