% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitLine.R
\name{unitLine}
\alias{unitLine}
\title{Add the Unit Line to a plot}
\usage{
unitLine(col = "blue", ...)
}
\arguments{
\item{col}{The colour of the line}

\item{...}{Used to pass any other parameters to the function \code{abline}}
}
\description{
A useful shortcut
}
\details{
This will simply draw the line y=x on any existing plot
}
\examples{
x <- rnorm(10)
y <- rnorm(10)
plot(x, y)
unitLine()

}
\seealso{
\code{\link{abline}}
}
