% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units2bed.R
\name{units2bed}
\alias{units2bed}
\title{Export a BED file(s) for a given set of units}
\usage{
units2bed(cdf, units, scoreCel = NULL, thresh = 8, singleBed = TRUE,
  upCol = "green", downCol = "red", ucscTracks = c("ensGene",
  "knownAlt", "refGene"), ...)
}
\arguments{
\item{cdf}{an \code{AffymetrixCdfFile}}

\item{units}{the units to be exported}

\item{scoreCel}{optional. An AffymetrixCelFile containing the group-level, (exon-level) scores for the given units.}

\item{thresh}{The value above which the scores are indicated using \code{upCol} & \code{downCol}}

\item{singleBed}{logical. If \code{TRUE}, all units will be output as a single BED file, 
otherwise each unit will be written as an individual BED file}

\item{upCol}{The colour to display for probesets with a positive score}

\item{downCol}{The colour to display for probesets with a negative score}

\item{ucscTracks}{The additional tracks to display in the UCSC browser}

\item{...}{not used}
}
\value{
Returns \code{TRUE} invisibly
  
Files will be written to the current working directory in the .BED format. 
If \code{singleBed=TRUE} a single file will be written including the current cdf name & any tags supplied via the scoreCel file. 
If \code{singleBed=FALSE} an individual .BED file will be written for each unit. 
The name will consist of any tags on the scoreCel file plus the unit ID as annotated on the cdf.
}
\description{
Enables easy uploading of probeset structure & AS events to a genome browser
}
\details{
This function enables easy display of the unit & group (probeset) structure of a given gene, or set of genes. 
In the case of custom generated CDF files this can be very useful for displaying the structure.
  
If scores are not provided the displayed tracks will default to black,
however if scores are provided, any probesets with score above the threshold will be displayed in 
colours of varying intensity, using those supplied in \code{upCol} & \code{downCol}.

All files with the genomic co-ordinates must also be present in the same directory as the supplied cdf. 
The name of this file must end with "mapping.txt" and must also contain the full cdf name somewhere in it's name.
}
\seealso{
\code{\link{AffymetrixCelFile}}, \code{\link{AffymetrixCdfFile}}
}
