% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.BMEA.R
\name{volcanoPlot.BMEA}
\alias{volcanoPlot.BMEA}
\title{Draw a volcano plot from a BMEA array}
\usage{
volcanoPlot.BMEA(data, plotVal = "mean", cont = 1, hLight = NULL,
  hLightCol = "blue", hLight.lab.cex = 0.7, hLight.pt.cex = 1,
  hLight.n = 10, pch = 16, cex = 0.35, main, ...)
}
\arguments{
\item{data}{the results from analysis with the BMEA model. 
Can be an \code{\link{AffymetrixCelSetList}}, \code{array} or \code{matrix}}

\item{plotVal}{determines which summary statistic to plot. 
Defaults to mean, but can also accept median or any other saved summary statistic named in the array}

\item{cont}{determines which contrast to plot. Equates to the the third dimension of any BmeaArray provided}

\item{hLight}{specify any points to be highlighted by unit ID}

\item{hLightCol}{the colour to use when highlighting points}

\item{hLight.lab.cex}{the scaling factor for point labelling}

\item{hLight.pt.cex}{the scaling factor to use when highlighting points}

\item{hLight.n}{the number of points to highlight}

\item{pch}{see \code{\link{plot}}}

\item{cex}{see \code{\link{plot}}}

\item{main}{see \code{\link{title}}}

\item{...}{any further parameters to be passed to \code{\link{plot}}}
}
\description{
Draws a volcano plot incorporating the possiblity of points with an infinite value on 
the y-axis.
}
\details{
Will draw a volcano plot from the results. Only a single contrast can be plotted, 
and any requested points will be highlighted, up to the maximum specified by \code{hLight.n}
}
\seealso{
\code{\link{extractBmeaArray}}
}

