% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBmeaBatch.R
\name{writeBmeaBatch}
\alias{writeBmeaBatch}
\title{Write the parameters to disk for a batch of genes}
\usage{
writeBmeaBatch(batchData, paramToWrite = c("c", "mu", "phi"), ...,
  nodePath = NULL, writeLog = TRUE, verbose = 0)
}
\arguments{
\item{batchData}{contains the results for a batch of genes after analysis with BMEA. 
Must be the output from \code{\link{fitBmeaBatch}}}

\item{paramToWrite}{the model parameters to write to disk. 
Any non-model parameters will be ignored}

\item{...}{required for passing arguments from higher level functions}

\item{nodePath}{the path to be used if being analysed on a specific node.}

\item{writeLog}{specifies whether a log file will be written for each batch of genes}

\item{verbose}{determines how much information will be displayed}
}
\value{
Returns a list with components:
\itemize{
\item{$units}{ the units that have just been written to disk}
\item{$parameters}{ the parameters for shich the units have been written}
\item{$time}{ the system time that the process was completed at, in POSIXct format}
}
}
\description{
Writes the quantiles & summary statistics to disk for a batch of genes
}
\details{
This function will write the requested model parameters to disk, 
along with any contrasts simulated, for a batch of genes. 
The \code{batchData} input must be the output from \code{fitBmeaBatch}.

If \code{nodePath=NULL}, model parameters will be written to the subdirectory \code{bmeaData/modelData/parentName,parameterName/chipName},
otherwise the output path will be written to \code{nodePath/bmeaData/modelData/parentName,parameterName/chipName}.

If contrasts have been supplied, the contrast data will be written to the subdirectories 
\code{bmeaData/contrastData/parentName,logFC/chipName} & \code{bmeaData/contrastData/parentName,phiLogFC/chipName}. 
If a \code{nodePath} has been supplied, this path will similarly appear at the beginning of the specified paths.

If \code{writeLog=TRUE}, a log file will be written to the directory \code{bmeaData} with the name \code{parentName.log.txt}. 
The units, parameters and time & date informastion are stored in this file.
}
\seealso{
\code{\link{fitBmeaBatch}}
}

