% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScore.R
\name{zScore}
\alias{zScore}
\title{Calculate the Z-Score for a given transcript/unit}
\usage{
zScore(PM, lambda, delta, exons = NULL)
}
\arguments{
\item{PM}{the raw observed intensities for the given unit. 
Must be a matrix with dimensions nrows=nProbes (K) & ncol= nChips (I).}

\item{lambda}{the vector of means for the log-normal distribution of the background signal.}

\item{delta}{the vector of standard deviations for the log-normal distribution of the background signal.}

\item{exons}{the exon structure of the gene as factors. 
If supplied, the function will return a value for each exon.}
}
\value{
A list with the components:
\itemize{
\item{$gene}{ the Z-Score for the entire gene}
\item{$exons}{ the Z-Score for each exon}
}

If the \code{exons} vector is not supplied, the \code{$exons} component will return \code{NULL}.
}
\description{
Based on the background priors for each probe, 
the Z-score is a measure of whether there is any detectable signal above background (DABG)
}
\details{
If no signal is present for a group of probes, any observed intensites will be samples 
from the distribution for background signal. 
Genes with z-score below a pre-defined threshold (e.g. 1.96) can be declared as not expressed, 
and likewise for any given exon.

If cacluating the value on a cumulative basis, individual scores will first be calculated for each exon, 
and then on an increasing basis.
}
\examples{
## Assuming the PM matrix, and the vectors of means & sds
## have already been obtained
PM <- matrix(exp(rnorm(120, 4, 2)), nrow=12, ncol=10) # A matrix of BG only pseudo-PM values
lambda <- matrix(4, nrow=12, ncol=10) # E[BG] at each probe
delta <- matrix(2, nrow=12, ncol=10) # SD[BG] at each probe
exons <- as.factor(rep(paste("Exon",1:3, sep=""), each=4)) # The exon structure
z <- zScore(PM, lambda, delta, exons) # The zScore
z

}

